refactor:(common/enum): creating index.ts file
