+++ packages/backend/src/common/enums/index.ts
@@ -1,0 +1,1 @@
+export * from './user.enum';
